requirements.txt
torch
torchvision
streamlit
timm==0.9.12
Pillow==10.0.1
numpy
scikit-learn
